[UnitTest] fix issue on test running and add a new test to check the access of incident logs comment #505
