+++ pytest.ini
@@ -4,3 +4,5 @@ python_files = tests.py test_*.py *_tests.py
 addopts = --create-db
 markers =
     django_db: marks tests as using the django database
+; testpaths = governanceplatform
+; pythonpath = governanceplatform
